requirements.txt
aiogram==3.0.0
python-dotenv==0.21.0
aiosqlite==0.19.0